clang-tidy: fix types

+++ tasks/seq/kholin_k_multidimensional_integrals_rectangle/include/ops_seq.hpp
@@ -33,7 +33,7 @@ class TestTaskSequential : public ppc::core::Task {
   size_t sz_upper_limits_;
 
   double Integrate(const Function& f, const std::vector<double>& l_limits, const std::vector<double>& u_limits,
-                   const std::vector<double>& h, std::vector<double>& f_values, double curr_index_dim, double dim,
+                   const std::vector<double>& h, std::vector<double>& f_values, int curr_index_dim, double dim,
                    double n);
   double IntegrateWithRectangleMethod(const Function& f, std::vector<double>& f_values,
                                       const std::vector<double>& l_limits, const std::vector<double>& u_limits,
